Improved demos, changed Docker image

+++ ffm/src/main/java/org/przybyl/ddj20/foreign/ForeignLinkerAPIDemo.java
@@ -27,8 +27,8 @@ public class ForeignLinkerAPIDemo {
 
     public static void main(String[] args) throws Throwable {
 
-        long uid = getUid();
-        System.out.println("Running with UID "+uid);
+        short uid = getUid();
+        System.out.println("Running with UID " + uid);
 
         if (isRoot(uid)) {
             System.err.println("Oh no, don't run me as root, PLEASE!!!11one");
@@ -38,7 +38,7 @@ public class ForeignLinkerAPIDemo {
 
     }
 
-    private static long getUid() throws Throwable {
+    private static short getUid() throws Throwable {
         // calling https://man7.org/linux/man-pages/man2/geteuid.2.html
         var nativeLinker = Linker.nativeLinker();
         var stdlib = nativeLinker.defaultLookup();
@@ -46,11 +46,11 @@ public class ForeignLinkerAPIDemo {
             .orElseThrow(() -> new IllegalStateException("Cannot find `getuid`"));
         var funcDesc = FunctionDescriptor.of(ValueLayout.JAVA_SHORT);
         var getuidHandle = nativeLinker.downcallHandle(getuid, funcDesc);
-        return (long) getuidHandle.invoke();
+        return (short) getuidHandle.invoke();
     }
 
-    private static boolean isRoot(long uid) {
-        return uid == 0L;
+    private static boolean isRoot(short uid) {
+        return uid == 0;
     }
